+++ include/bgfx/c99/bgfx.h
@@ -2390,7 +2390,7 @@ BGFX_C_API void bgfx_encoder_set_condition(bgfx_encoder_t* _this, bgfx_occlusion
  *  _fstencil is applied to both front and back facing primitives.
  *
  */
-BGFX_C_API void bgfx_encoder_set_stencil(bgfx_encoder_t* _this, uint32_t _fstencil, uint32_t _bstencil);
+BGFX_C_API void bgfx_encoder_set_stencil(bgfx_encoder_t* _this, uint64_t _fstencil, uint64_t _bstencil);
 
 /**
  * Set scissor for draw primitive.
@@ -2941,7 +2941,7 @@ BGFX_C_API void bgfx_set_condition(bgfx_occlusion_query_handle_t _handle, bool _
  *  _fstencil is applied to both front and back facing primitives.
  *
  */
-BGFX_C_API void bgfx_set_stencil(uint32_t _fstencil, uint32_t _bstencil);
+BGFX_C_API void bgfx_set_stencil(uint64_t _fstencil, uint64_t _bstencil);
 
 /**
  * Set scissor for draw primitive.
@@ -3672,7 +3672,7 @@ struct bgfx_interface_vtbl
     void (*encoder_set_marker)(bgfx_encoder_t* _this, const char* _marker);
     void (*encoder_set_state)(bgfx_encoder_t* _this, uint64_t _state, uint32_t _rgba);
     void (*encoder_set_condition)(bgfx_encoder_t* _this, bgfx_occlusion_query_handle_t _handle, bool _visible);
-    void (*encoder_set_stencil)(bgfx_encoder_t* _this, uint32_t _fstencil, uint32_t _bstencil);
+    void (*encoder_set_stencil)(bgfx_encoder_t* _this, uint64_t _fstencil, uint64_t _bstencil);
     uint16_t (*encoder_set_scissor)(bgfx_encoder_t* _this, uint16_t _x, uint16_t _y, uint16_t _width, uint16_t _height);
     void (*encoder_set_scissor_cached)(bgfx_encoder_t* _this, uint16_t _cache);
     uint32_t (*encoder_set_transform)(bgfx_encoder_t* _this, const void* _mtx, uint16_t _num);
@@ -3717,7 +3717,7 @@ struct bgfx_interface_vtbl
     void (*set_marker)(const char* _marker);
     void (*set_state)(uint64_t _state, uint32_t _rgba);
     void (*set_condition)(bgfx_occlusion_query_handle_t _handle, bool _visible);
-    void (*set_stencil)(uint32_t _fstencil, uint32_t _bstencil);
+    void (*set_stencil)(uint64_t _fstencil, uint64_t _bstencil);
     uint16_t (*set_scissor)(uint16_t _x, uint16_t _y, uint16_t _width, uint16_t _height);
     void (*set_scissor_cached)(uint16_t _cache);
     uint32_t (*set_transform)(const void* _mtx, uint16_t _num);
